Added crude js support with some examples
This is awful code, copy pasted all over the place, changes yaml files to json because I couldn't figure out parse yaml in js. This
adds dependencies on new libraries that may or may not be desired, is not backwards compatible and also breaks compatibility with java.

However, it does run. Very slowly.